Fix unsigned-signed comparisons that GCC 12.2 warns about.

+++ tools/ld-process-ac3/decode/AC3Framer.hpp
@@ -99,7 +99,7 @@ struct AC3Framer {
             if (currentAc3Size != 0) { // append to the ac3 block
                 // ac3Buffer.emplace_back(buffer[i]); // todo
                 ac3Buffer.insert(ac3Buffer.end(), buffer[bytePosition]);
-                if (ac3Buffer.size() == currentAc3Size) {
+                if (static_cast<int>(ac3Buffer.size()) == currentAc3Size) {
                     currentAc3Size = 0;
                     // assert(ac3Buffer[4] == 0x1c); // fscod and frmsizecod fixed until lookup above implemented
                     return ac3Buffer;
